refactor: modal 관련 상태 값 및 함수 useModal hook으로 분리
