+++ src/hooks/useModal.js
@@ -1,0 +1,21 @@
+import { useState } from 'react';
+
+const useModal = () => {
+  const [isOpenModal, setIsOpenModal] = useState(false);
+
+  const openModal = () => {
+    setIsOpenModal(true);
+  };
+
+  const closeModal = () => {
+    setIsOpenModal(false);
+  };
+
+  return {
+    isOpenModal,
+    openModal,
+    closeModal,
+  };
+};
+
+export default useModal;
